requirements.txt
attrs==20.3.0
Flask==1.1.2
pandas==1.1.5
pdftotext==2.1.5
Pillow==8.1.0
python-docx==0.8.10
requests==2.25.1